# openapi.yaml
openapi: 3.0.3
info:
  title: PBI Enterprise API
  version: 1.0.0
  description: |
    Presence-Bound Identity (PBI) enterprise API:
    - issues action/artifact-bound challenges
    - verifies WebAuthn assertion bundles (UP+UV required)
    - emits receipts
    - meters usage and exposes invoices
servers:
  - url: https://api.kojib.com
  - url: http://localhost:8080
components:
  securitySchemes:
    ApiKeyBearer:
      type: http
      scheme: bearer
      bearerFormat: API_KEY
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    PBIChallenge:
      type: object
      properties:
        id: { type: string, format: uuid }
        challengeB64Url: { type: string }
        purpose:
          type: string
          enum: [ACTION_COMMIT, ARTIFACT_AUTHORSHIP, EVIDENCE_SUBMIT, ADMIN_DANGEROUS_OP]
        actionHashHex: { type: string, pattern: "^[0-9a-f]{64}$" }
        expiresAtIso: { type: string, format: date-time }
      required: [id, challengeB64Url, purpose, actionHashHex, expiresAtIso]
    ChallengeRequest:
      type: object
      properties:
        purpose:
          type: string
          enum: [ACTION_COMMIT, ARTIFACT_AUTHORSHIP, EVIDENCE_SUBMIT, ADMIN_DANGEROUS_OP]
        actionHashHex:
          type: string
          pattern: "^[0-9a-f]{64}$"
        ttlSeconds:
          type: integer
          minimum: 10
          maximum: 600
          default: 120
      required: [purpose, actionHashHex]
    ChallengeResponse:
      type: object
      properties:
        challenge:
          $ref: "#/components/schemas/PBIChallenge"
      required: [challenge]
    WebAuthnAssertionBundle:
      type: object
      properties:
        authenticatorDataB64Url: { type: string }
        clientDataJSONB64Url: { type: string }
        signatureB64Url: { type: string }
        credIdB64Url: { type: string }
        pubKeyPem: { type: string, description: "PEM encoded P-256 public key for verification (enterprise mode without enrollment)" }
      required: [authenticatorDataB64Url, clientDataJSONB64Url, signatureB64Url, credIdB64Url, pubKeyPem]
    VerifyRequest:
      type: object
      properties:
        challengeId: { type: string, format: uuid }
        assertion:
          $ref: "#/components/schemas/WebAuthnAssertionBundle"
      required: [challengeId, assertion]
    VerifyResponse:
      type: object
      properties:
        ok: { type: boolean }
        decision:
          type: string
          enum: [PBI_VERIFIED, FAILED, EXPIRED, REPLAYED]
        reason:
          type: string
          nullable: true
        receiptId: { type: string, format: uuid, nullable: true }
        receiptHashHex: { type: string, nullable: true }
        challenge:
          type: object
          properties:
            id: { type: string, format: uuid }
            purpose:
              type: string
              enum: [ACTION_COMMIT, ARTIFACT_AUTHORSHIP, EVIDENCE_SUBMIT, ADMIN_DANGEROUS_OP]
            actionHashHex: { type: string, pattern: "^[0-9a-f]{64}$" }
          required: [id, purpose, actionHashHex]
      required: [ok, decision]
...
    Receipt:
      type: object
      properties:
        id: { type: string, format: uuid }
        challengeId: { type: string, format: uuid }
        receiptHashHex: { type: string }
        decision:
          type: string
          enum: [PBI_VERIFIED, FAILED, EXPIRED, REPLAYED]
        createdAt: { type: string, format: date-time }
      required: [id, challengeId, receiptHashHex, decision, createdAt]
    ReceiptResponse:
      type: object
      properties:
        receipt:
          $ref: "#/components/schemas/Receipt"
      required: [receipt]
...
security:
  - ApiKeyBearer: []
paths:
...
  /v1/pbi/challenge:
    post:
      summary: Create a PBI challenge bound to an action/artifact hash
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChallengeRequest" }
      responses:
        "200":
          description: Challenge created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChallengeResponse" }
        "401":
          description: Missing API key
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Invalid API key
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /v1/pbi/verify:
    post:
      summary: Verify a WebAuthn assertion against a stored challenge and issue a receipt
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VerifyRequest" }
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VerifyResponse" }
        "400":
          description: Failed / expired / replayed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VerifyResponse" }
  /v1/pbi/receipts/{receiptId}:
    get:
      summary: Fetch a receipt by id
      parameters:
        - in: path
          name: receiptId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Receipt
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReceiptResponse" }
        "404":
          description: Receipt not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...